server.js now initializes mongodb collections!
At startup server.js now attempts to create the necessary collections.
If they exist already they will NOT be overwritten. If default data
population is necessary, a JSON object may be passed for population.

Unless I am missing something, this means the users will NOT have to
manually add the ‘settings’ collection. It will happen automatically in
the background.

I also set it up to initialize the Entries collection as ‘entries’ if
no name has been provided, and the database does not exist. No default
data population occurs for entries.

This requires that the website be setup with a mongo user with write
permissions. Of course we were going to need that for the API anyway.