introduce express.Router( ), small renaming
Rename a few things here and there to help clarify the situation.

This change introduces use of express.Router( ), which provides the routing
functions of express without the full app.
This also shows an experiment observing the interaction betwen express( ) and
express.Router( ).
There are a few reasons for doing this:

Using express.Router( ) allows introspecting existing routes.
As a REST application, our content payloads should include URLs, links to other
resources available from our API.  As a bonus feature, it's often nice for a
meta page to describe the links, status, and version available from this
particular API.  By far, the easiest way to implement this is to introspect the
existing routes.

Separating the code makes testing easier.
This lays the ground work for breaking apart the code into smaller, more
testable modules.