Cleaned up template default values for deployment
In this checkin I've removed the parameters (DISPLAY_UNITS,
MONGO_SETTINGS_COLLECTION, MONGO_TREATMENTS_COLLECTION,
MONGO_DEVICESTATUS_COLLECTION, ENABLE, API_SECRET) from the Deployment
UI Screen.  However I am still creating the environment variables with
the defaults on deployment.

I've done this so we can enforce the connection string is required and
prevent deployment until it is filled in.  By leaving all of the other
variables in place it was confusing to a user which value was preventing
progress.

I think this provides a cleaner deployment experience for users, while
still giving them the ability to modify the other variables post
deployment in the Azure Portal.