Bulgarian language  edit - 17.08.2015
Hi , i changed some of the translation and added announcement to the list .  
Is there another language file with all the words from the treatment profile?
Is it going to be possible to have more than one treatment profiles and change to the stored setting when that occur.  For example when you are sick. 
If there is anything else i am ready to help