Build docker images with Travis
- Add Docker hub secrets to Travis
- Push to Docker Hub
- Ignore .git directory and things like coverage logs, so as to minimise
  the container size

I placed the Docker hub login commands in the commands in the .travis.yml file,
rather than in the Makefile. This is because the Makefile would display the
Docker password to stdout, which would then become publically available in the
Travis logs.

I also run `git clean -fd .` and `git checkout -- .` to ensure that anything
written to the repo directory during tests isn't added to the docker container.
I do this in the Travis file rather than the Makefile, since running things with
`make` reverted my changes while experimenting, and I want to save others the
same issue.

To make this work on Travis, we need to set secrets for the DOCKER_USER and
DOCKER_PASS environment variable.