Generating codacy coverage should not fail build
This should force the coverage job to return with a success error code.
Codacy changed support for legacy projects.  This change should not cause
builds to fail and should not influence ability to merge pull requests.  Until
further work with coverage with codacy, this change emits a warning of "NO
COVERAGE" while allowing the build to pass, rather than failing the build.