ensure obfuscation only affects rendered values
The results to render may come from an in-memory cache, or from the database.
If they came from the in-memory cache, it's possible that obscurring the device
field could actually change the values in memory.  When obscurring, the device
field, this patch copies all of the values in order to ensure that the
in-memory copies are unaffected.  With this patch, cached objects should not be
affected.