redefine bundles, refactor views to use assets
This patch allows splitting the bundles for the biggest vendored dependencies
and for Nightscout's core application code.  The result changes te bundle from
being a single asset to a several assets.
For app:
  * "js/bundle.runtime.js"
  * "js/bundle.vendors.js"
  * "js/bundle.vendor-jquery.js"
  * "js/bundle.vendor-moment.js"
  * "js/bundle.vendor-d3.js"
  * "js/bundle.app.js"

Clock is similar:
  * "js/bundle.runtime.js"
  * "js/bundle.vendors.js"
  * "js/bundle.vendor-jquery.js"
  * "js/bundle.clock.js"

In addition, a file is kept, including the relevant assets per entrypoint.
This file is used by Nightscout to generate the list of bundled assets to
include per page automatically.

There is a modest size decrease of ~20% for the bundle sizes, mostly by
pointing to jquery and d3's minimized assets.